requirements.txt
pandas
openpyxl
sentence-transformers
torch
psycopg2-binary
python-dotenv
fastapi
uvicorn[standard]
email-validator
